Create get_binned_cube function
Add a function get_binned_cube which creates a binned version of the initial data cube for the user to work with as well as the new dimensions of the binned cube. This function is heavily based on the current bin_cube function but allows the user to retreive the binned cube for later analysis steps (e.g. masking...)