Updated plotting and fixed amplitude initial guess
The initial guess from the amplitude was not continuum subtracted which was leading to wrong results. This has been fixed.

We also now are plotting a bit differently.  have updated this by, when plotting, taking the center point as the closest axis value. The way to visually fix this is to not use the true center of the peak for the plotting but to instead use the axis value that is closest to this true value.