Added fit_WVT function
Added fit_WVT function which performs a WVT algorithm on a specific region passed by the user. The bins are then fitted individually with fit_spectrum_region() and the velocity values produced by the fit are then associated to all pixels of the bin to create a velocity map. This function is highly preliminary and updates will be needed to make it work appropriately.