Updated WVT algorithm
Updated the functions allowing the user to perform a WVT procedure on a specific region of the data. The created WVT bins can then be used to perform a fitting procedure and create velocity, broadening, amplitude and flux maps.

To Do : There seem to be an error in the fitting procedure while passing through all the bins. The resulting maps are missing some bins in a weird way...